Fix Add auth to admin
