+++ packages/admin/.env.example
@@ -1,0 +1,1 @@
+REACT_APP_API_URL=http://127.0.0.1:3333
