+++ packages/admin/package.json
@@ -8,6 +8,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "dotenv": "^8.2.0",
     "prop-types": "^15.7.2",
     "query-string": "^6.11.1",
     "ra-data-json-server": "^3.3.1",
